Hemodynamic Response Function with Half-Cosine Basis

This function models a hemodynamic response function (HRF) using four half-period cosine basis functions. The HRF consists of an initial dip, a rise to peak, a fall and undershoot, and a recovery to the baseline.

Usage

hrf_half_cosine(t, h1 = 1, h2 = 5, h3 = 7, h4 = 7, f1 = 0, f2 = 0)

Arguments

t

A vector of time values.

h1

Duration of the initial dip in seconds.

h2

Duration of the rise to peak in seconds.

h3

Duration of the fall and undershoot in seconds.

h4

Duration of the recovery to baseline in seconds.

f1

Height of the starting point.

f2

Height of the end point.

Value

A vector of HRF values corresponding to the input time values.

References

Woolrich, M. W., Behrens, T. E., & Smith, S. M. (2004). Constrained linear basis sets for HRF modelling using Variational Bayes. NeuroImage, 21(4), 1748-1761.