hrf_inv_logit

A hemodynamic response function using the difference of two Inverse Logit functions.

Usage

hrf_inv_logit(t, mu1 = 6, s1 = 1, mu2 = 16, s2 = 1, lag = 0)

Arguments

t

A vector of times.

mu1

The time-to-peak for the rising phase (mean of the first logistic function).

s1

The width (slope) of the first logistic function.

mu2

The time-to-peak for the falling phase (mean of the second logistic function).

s2

The width (slope) of the second logistic function.

lag

The time delay (default: 0).

Value

A vector of the difference of two Inverse Logit HRF values.

See also

Examples

hrf_inv_logit_basis <- hrf_inv_logit(seq(0, 20, by = 0.5), mu1 = 6, s1 = 1, mu2 = 16, s2 = 1)