hrf_sine

A hemodynamic response function using the Sine Basis Set.

Usage

hrf_sine(t, span = 24, N = 5)

Arguments

t

A vector of times.

span

The temporal window over which the basis sets span (default: 24).

N

The number of basis functions (default: 5).

Value

A matrix of sine basis functions.

See also

Examples

hrf_sine_basis <- hrf_sine(seq(0, 20, by = 0.5), N = 4)