HRF (hemodynamic response function) as a linear function of time

The `hrf_time` function computes the value of an HRF, which is a simple linear function of time `t`, when `t` is greater than 0 and less than `maxt`.

Usage

hrf_time(t, maxt = 22)

Arguments

t

A numeric value representing time in seconds.

maxt

A numeric value representing the maximum time point in the domain. Default value is 22.

Value

A numeric value representing the value of the HRF at the given time `t`.

See also

Examples

# Compute the HRF value for t = 5 seconds with the default maximum time
hrf_val <- hrf_time(5)

# Compute the HRF value for t = 5 seconds with a custom maximum time of 30 seconds
hrf_val_custom_maxt <- hrf_time(5, maxt = 30)